% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{can_snapshot}
\alias{can_snapshot}
\title{Check if Snapshot Capability is Available}
\usage{
can_snapshot(simulation_result)
}
\arguments{
\item{simulation_result}{Result object from twolife_simulation}
}
\value{
Logical indicating if snapshot_at_time() can be used
}
\description{
Helper function to check if a result has sufficient history detail
for snapshot_at_time() function.
}
\examples{
\dontrun{
result <- twolife_simulation(..., history_detail = "minimal")
if (can_snapshot(result)) {
  snapshot_at_time(result, 50)
} else {
  message("Re-run with history_detail='standard' or 'full'")
}
}

}
