% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.twolife_result}
\alias{plot.twolife_result}
\title{Plot Method for TWoLife Results}
\usage{
\method{plot}{twolife_result}(x, ...)
}
\arguments{
\item{x}{A 'twolife_result' object from \code{\link{twolife_simulation}}}

\item{...}{Additional arguments passed to \code{\link{plot_simulation_on_landscape}}.
Common options include: point_size, point_color, color_by, landscape_colors, main}
}
\value{
The input object \code{x}, invisibly
}
\description{
Generic plot method for twolife_result objects. Calls 
\code{\link{plot_simulation_on_landscape}} with convenient syntax.
}
\examples{
landscape <- create_fractal_landscape(
  cells_per_row = 15,
  fractality = 0.5,
  habitat_proportion = 0.4,
  return_as_landscape_params = TRUE
)

result <- twolife_simulation(
  landscape_params = landscape,
  individual_params = list(initial_population_size = 30),
  simulation_params = list(max_events = 1000),
  master_seed = 789
)

# Simple plot using S3 method
plot(result)

# With options
plot(result, point_size = 3, color_by = "genotype")

# Equivalent to:
plot_simulation_on_landscape(result, point_size = 3, color_by = "genotype")

}
\seealso{
\code{\link{plot_simulation_on_landscape}} for full documentation of plotting options
}
