% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_twolife_simulation}
\alias{run_twolife_simulation}
\title{Internal C++ Simulation Function}
\arguments{
\item{neighbor_radius}{Radius of density dependence influence}

\item{initial_population_size}{Number of individuals at start}

\item{vision_angle}{Angle used for orientation when dispersing}

\item{step_length}{Distance of dispersal event}

\item{base_dispersal_rate}{Rate at which individuals disperse}

\item{base_birth_rate}{Basal birth rate}

\item{base_mortality_rate}{Basal death rate}

\item{birth_density_slope}{Slope of birth density dependence}

\item{mortality_density_slope}{Slope of death density dependence}

\item{habitat}{Square matrix of environmental values (0=matrix, 1=habitat)}

\item{cell_size}{Resolution length of pixel side}

\item{density_type}{Density type (0=global, 1=local)}

\item{matrix_mortality_multiplier}{Mortality multiplier for matrix}

\item{matrix_dispersal_multiplier}{Dispersal multiplier for matrix}

\item{initial_placement_mode}{Initial position mode}

\item{boundary_condition}{Boundary condition type}

\item{max_events}{Maximum number of events}

\item{initial_x_coordinates}{X coordinates of initial individuals}

\item{initial_y_coordinates}{Y coordinates of initial individuals}

\item{genotype_means}{Genotypical trait means}

\item{genotype_sds}{Standard deviations}

\item{mutation_rates}{Mutation rates}

\item{plasticities}{Plasticities}

\item{sampling_points}{Number of sampling points}

\item{neutral_mode}{Neutral simulation mode}

\item{output_file}{Optional output file path}
}
\value{
List containing simulation results
}
\description{
Low-level C++ function for running TWoLife simulations. This function is for
internal use only - users should use \code{\link{twolife_simulation}} instead.
}
\seealso{
\code{\link{twolife_simulation}} for the user-friendly interface
}
\keyword{internal}
