% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{single_large}
\alias{single_large}
\title{Example landscape file: single large habitat patch}
\format{
dataset
}
\usage{
single_large
}
\description{
A pre-computed landscape with one large habitat patch
for demonstration purposes.
}
\keyword{datasets}
