% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{batch_validate_habitat_matching}
\alias{batch_validate_habitat_matching}
\title{Batch Validate Multiple Simulations}
\usage{
batch_validate_habitat_matching(
  result_list,
  point_size = 1.5,
  landscape_colors = "terrain"
)
}
\description{
Batch Validate Multiple Simulations
}
