% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_genotype_habitat_mismatch}
\alias{calculate_genotype_habitat_mismatch}
\title{Calculate Genotype-Habitat Mismatch Statistics Using Fitness}
\usage{
calculate_genotype_habitat_mismatch(
  simulation_result,
  return_individuals = FALSE
)
}
\description{
Calculate Genotype-Habitat Mismatch Statistics Using Fitness
}
