% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.genotype_habitat_mismatch}
\alias{print.genotype_habitat_mismatch}
\title{Print Method for Mismatch Statistics}
\usage{
\method{print}{genotype_habitat_mismatch}(x, ...)
}
\arguments{
\item{x}{A genotype_habitat_mismatch object}

\item{...}{Additional arguments (ignored)}
}
\value{
x (invisibly)
}
\description{
Print Method for Mismatch Statistics
}
