% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{quick_plot_result}
\alias{quick_plot_result}
\title{Quick Plot of Simulation Results}
\usage{
quick_plot_result(simulation_result, ...)
}
\arguments{
\item{simulation_result}{A twolife_result object}

\item{...}{Additional arguments passed to \code{\link{plot_simulation_on_landscape}}}
}
\description{
Simplified function for quickly visualizing TWoLife simulation results.
}
\examples{
result <- twolife_simulation(...)  # your simulation
quick_plot_result(result)

}
\seealso{
\code{\link{plot_simulation_on_landscape}} for full customization
}
