% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{several_small}
\alias{several_small}
\title{Example landscape file: several small habitat patches}
\format{
dataset
}
\usage{
several_small
}
\description{
A pre-computed landscape with multiple small habitat fragments
for demonstration purposes.
}
\keyword{datasets}
